+++ test/jobs/ICounter.sol
@@ -4,4 +4,5 @@ pragma solidity ^0.8.13;
 interface ICounter {
   function current() external view returns (uint256);
   function myResolver(string calldata pass) external view returns (bool, bytes memory);
+  function getLastExecuteByJobKey() external view returns (bytes32);
 }